+++ Sources/Sentry/SentryOptions.m
@@ -138,9 +138,6 @@ NSString *const kSentryDefaultEnvironment = @"production";
         self.swiftAsyncStacktraces = NO;
         self.enableSpotlight = NO;
         self.spotlightUrl = @"http://localhost:8969/stream";
-        self.enableMetrics = NO;
-        self.enableDefaultTagsForMetrics = YES;
-        self.enableSpanLocalMetricAggregation = YES;
 
 #if TARGET_OS_OSX
         NSString *dsn = [[[NSProcessInfo processInfo] environment] objectForKey:@"SENTRY_DSN"];
@@ -544,18 +541,6 @@ NSString *const kSentryDefaultEnvironment = @"production";
 
     if ([options[@"spotlightUrl"] isKindOfClass:[NSString class]]) {
         self.spotlightUrl = options[@"spotlightUrl"];
-    }
-
-    [self setBool:options[@"enableMetrics"] block:^(BOOL value) { self->_enableMetrics = value; }];
-
-    [self setBool:options[@"enableDefaultTagsForMetrics"]
-            block:^(BOOL value) { self->_enableDefaultTagsForMetrics = value; }];
-
-    [self setBool:options[@"enableSpanLocalMetricAggregation"]
-            block:^(BOOL value) { self->_enableSpanLocalMetricAggregation = value; }];
-
-    if ([self isBlock:options[@"beforeEmitMetric"]]) {
-        self.beforeEmitMetric = options[@"beforeEmitMetric"];
     }
 
     if ([options[@"experimental"] isKindOfClass:NSDictionary.class]) {