+++ Sources/Sentry/include/SentryPrivate.h
@@ -2,7 +2,6 @@
 #import "SentryDispatchQueueWrapper.h"
 #import "SentryNSDataUtils.h"
 #import "SentryRandom.h"
-#import "SentryStatsdClient.h"
 #import "SentryTime.h"
 
 // Headers that also import SentryDefines should be at the end of this list